Suite1.1 Container model changes
SVN-Revision: 13219

+++ WEB-INF/src/edu/wustl/catissuecore/bizlogic/SpecimenEventParametersBizLogic.java
@@ -257,7 +257,10 @@ public class SpecimenEventParametersBizLogic extends DefaultBizLogic
 					 */
 					//Specimen proxySpecimen = (Specimen)HibernateMetaData.getProxyObjectImpl(specimen);
 					dao.update(specimen, sessionDataBean, true, true, false);
-					dao.delete(prevPosition);	
+					if(prevPosition!=null)
+					{
+						dao.delete(prevPosition);
+					}	
 					try
 					{
 						CatissueCoreCacheManager catissueCoreCacheManager = CatissueCoreCacheManager.getInstance();
